+++ refactor/src/interfaces/PlayerConfig.ts
@@ -1,48 +1,4 @@
-export interface Context {
-  mode: string;
-  threshold?: number;
-  authToken?: string;
-  sid: string;
-  did: string;
-  uid: string;
-  channel: string;
-  pdata: Pdata;
-  contextRollup: ContextRollup;
-  tags: string[];
-  cdata?: Cdata[];
-  timeDiff?: number;
-  objectRollup?: ObjectRollup;
-  host?: string;
-  endpoint?: string;
-  userData?: {
-    firstName: string;
-    lastName: string;
-  };
-}
-export interface Pdata {
-  id: string;
-  pid: string;
-  ver: string;
-}
-
-export interface ContextRollup {
-  l1?: string;
-  l2?: string;
-  l3?: string;
-  l4?: string;
-}
-
-export interface Cdata {
-  type: string;
-  id: string;
-}
-
-export interface ObjectRollup {
-  l1?: string;
-  l2?: string;
-  l3?: string;
-  l4?: string;
-}
+import { Context } from "./Telemetry";
 
 export interface Config {
   traceId?: string;
@@ -60,7 +16,7 @@ export interface Config {
 }
 
 export interface QumlPlayerConfig {
-  context: Context;
+  context: Context; //TODO: Query - why is the telemetry context defined here?
   metadata: any; //TODO: define metadata interface
   data: any; //TODO: define data interface
   config: Config;
